+++ internal/usecase/room.go
@@ -6,72 +6,13 @@ import (
 	"slices"
 	"sync"
 
-	log "github.com/sirupsen/logrus"
-
 	"dishdash.ru/internal/domain"
+	"dishdash.ru/internal/usecase/event"
 	"dishdash.ru/internal/usecase/state"
-	"dishdash.ru/pkg/algo"
+	algo "dishdash.ru/pkg/algo"
+	log "github.com/sirupsen/logrus"
 	orderedmap "github.com/wk8/go-ordered-map/v2"
 )
-
-type State string
-
-type VoteType string
-
-const (
-	VoteTypeMatch  VoteType = "match"
-	VoteTypeFinish VoteType = "finish"
-)
-
-type Match struct {
-	ID    int           `json:"id"`
-	Place *domain.Place `json:"card"`
-}
-
-type OptionID int64
-
-type VoteOption struct {
-	ID   OptionID `json:"id"`
-	Desc string   `json:"description"`
-}
-
-type Vote interface {
-	isVote()
-}
-
-type BaseVote struct {
-	ID      int64        `json:"id"`
-	Options []VoteOption `json:"options"`
-	Type    VoteType     `json:"type"`
-	votes   map[string]OptionID
-}
-
-type FinishVote struct {
-	BaseVote
-}
-
-func (v *FinishVote) isVote() {}
-
-type MatchVote struct {
-	BaseVote
-	Place *domain.Place `json:"card"`
-}
-
-func (v *MatchVote) isVote() {}
-
-const (
-	OptionIDLike OptionID = iota
-	OptionIDDislike
-
-	OptionIDFinish
-	OptionIDContinue
-)
-
-type VoteResult struct {
-	Type     VoteType `json:"type"`
-	VoteID   int64    `json:"voteId"`
-	OptionID OptionID `json:"optionId"`
-}
 
 type Room struct {
 	id   string
@@ -86,10 +27,10 @@ type Room struct {
 	usersPlace map[string]*domain.Place
 	places     []*domain.Place
 	swipes     []*domain.Swipe
-	matches    []*Match
+	matches    []*event.Match
 	result     *domain.Place
 
-	votes map[int64]Vote
+	votes map[int64]event.VoteAnnounce
 
 	lobbyUseCase     Lobby
 	placeUseCase     Place
@@ -110,31 +51,34 @@ func NewRoom(
 	r := &Room{
 		id:               lobby.ID,
 		lobby:            lobby,
-		usersMap:         make(map[string]*domain.User),
-		places:           make([]*domain.Place, 0),
-		usersPlace:       make(map[string]*domain.Place),
-		swipes:           make([]*domain.Swipe, 0),
-		matches:          make([]*Match, 0),
-		votes:            make(map[int64]Vote),
 		lobbyUseCase:     lobbyUseCase,
 		placeUseCase:     placeUseCase,
 		swipeUseCase:     swipeUseCase,
 		userUseCase:      userUseCase,
 		placeRecommender: placeRecommender,
-		log:              log.WithFields(log.Fields{"room": lobby.ID}),
+
+		usersMap:   make(map[string]*domain.User),
+		places:     make([]*domain.Place, 0),
+		usersPlace: make(map[string]*domain.Place),
+		swipes:     make([]*domain.Swipe, 0),
+		matches:    make([]*event.Match, 0),
+		votes:      make(map[int64]event.VoteAnnounce),
+		log:        log.WithFields(log.Fields{"room": lobby.ID}),
 	}
 
-	r.votes[0] = &FinishVote{
-		BaseVote{
-			ID:      0,
-			Options: []VoteOption{{ID: OptionIDFinish, Desc: "Finish"}, {ID: OptionIDContinue, Desc: "Continue"}},
-			Type:    VoteTypeFinish,
-			votes:   make(map[string]OptionID),
+	r.votes[0] = event.FinishVote{
+		BaseVote: event.BaseVote{
+			ID: 0,
+			Options: []event.VoteOption{
+				{ID: event.OptionIDFinish, Desc: "Finish"},
+				{ID: event.OptionIDContinue, Desc: "Continue"},
+			},
+			Type:  event.VoteTypeFinish,
+			Votes: make(map[string]event.OptionID),
 		},
 	}
 
 	r.state = lobby.State
-	r.log.Debugf("state: %s", r.state)
 
 	if r.state != domain.InLobby {
 		// can't load swiping lobby
@@ -144,11 +88,11 @@ func NewRoom(
 		}
 	}
 
-	if r.state == domain.Finished || r.state == domain.InLobby {
-		err := r.loadFromDB()
-		if err != nil {
-			return nil, fmt.Errorf("failed to load from db: %w", err)
-		}
+	r.log.Debugf("state: %s", r.state)
+
+	err := r.Load()
+	if err != nil {
+		return nil, fmt.Errorf("failed to load room: %w", err)
 	}
 
 	return r, nil
@@ -158,7 +102,24 @@ func (r *Room) ID() string {
 	return r.id
 }
 
-func (r *Room) loadFromDB() error {
+func (r *Room) Users() []*domain.User {
+	r.lock.RLock()
+	defer r.lock.RUnlock()
+	return r.users()
+}
+
+func (r *Room) users() []*domain.User {
+	users := make([]*domain.User, 0)
+	for _, user := range r.usersMap {
+		users = append(users, user)
+	}
+	return users
+}
+
+func (r *Room) Load() error {
+	r.lock.Lock()
+	defer r.lock.Unlock()
+
 	var err error
 	r.swipes, err = r.swipeUseCase.GetSwipesByLobbyID(context.Background(), r.id)
 	if err != nil {
@@ -181,7 +142,7 @@ func (r *Room) loadFromDB() error {
 	return nil
 }
 
-func (r *Room) mathesFromSwipes() ([]*Match, error) {
+func (r *Room) mathesFromSwipes() ([]*event.Match, error) {
 	r.log.WithFields(log.Fields{"swipeCount": len(r.swipes)}).Debug("restore matches from db")
 	swipeCount := orderedmap.New[int64, int]()
 
@@ -196,7 +157,7 @@ func (r *Room) mathesFromSwipes() ([]*Match, error) {
 		swipeCount.Set(swipe.PlaceID, c+1)
 	}
 
-	matches := make([]*Match, 0)
+	matches := make([]*event.Match, 0)
 
 	for e := swipeCount.Oldest(); e != nil; e = e.Next() {
 		r.log.WithFields(log.Fields{"place": e.Key, "count": e.Value}).Debug("swipe count")
@@ -205,7 +166,7 @@ func (r *Room) mathesFromSwipes() ([]*Match, error) {
 			if err != nil {
 				return nil, fmt.Errorf("failed to get place: %w", err)
 			}
-			matches = append(matches, &Match{
+			matches = append(matches, &event.Match{
 				ID:    len(matches),
 				Place: place,
 			})
@@ -215,79 +176,101 @@ func (r *Room) mathesFromSwipes() ([]*Match, error) {
 	return matches, nil
 }
 
-func (r *Room) GetNextPlaceForUser(id string) *domain.Place {
-	r.lock.RLock()
-	defer r.lock.RUnlock()
-	return r.usersPlace[id]
-}
-
-func (r *Room) Users() []*domain.User {
-	r.lock.RLock()
-	defer r.lock.RUnlock()
-	return r.users()
-}
-
-func (r *Room) users() []*domain.User {
-	users := make([]*domain.User, 0)
-	for _, user := range r.usersMap {
-		users = append(users, user)
+func (r *Room) setState(state domain.LobbyState) error {
+	r.state = state
+	err := r.lobbyUseCase.SetLobbyState(context.Background(), r.lobby.ID, state)
+	if err != nil {
+		return fmt.Errorf("failed to set lobby state: %w", err)
 	}
-	return users
-}
-
-func (r *Room) Settings() UpdateLobbySettingsInput {
-	r.lock.RLock()
-	defer r.lock.RUnlock()
-	return r.settings()
-}
-
-func (r *Room) settings() UpdateLobbySettingsInput {
-	return UpdateLobbySettingsInput{
-		ID:       r.id,
-		PriceAvg: r.lobby.PriceAvg,
-		Location: r.lobby.Location,
-		Tags: algo.Map(r.lobby.Tags, func(t *domain.Tag) int64 {
-			return t.ID
-		}),
-	}
+	return nil
 }
 
 func (r *Room) OnJoin(c *state.Context[*Room]) error {
-	r.usersMap[c.User.ID] = c.User
+	r.lock.Lock()
+	defer r.lock.Unlock()
 
-	// c.BroadcastToOthers(event.UserJoined, event.UserJoinedEvent{
-	// 	ID:     c.User.ID,
-	// 	Name:   c.User.Name,
-	// 	Avatar: c.User.Avatar,
-	// })
+	switch r.state {
+	case domain.InLobby:
+		r.usersMap[c.User.ID] = c.User
+
+		err := r.syncUsersWithBd(c.Ctx)
+		if err != nil {
+			return fmt.Errorf("failed to sync users with db: %w", err)
+		}
+
+		c.BroadcastToOthers(event.UserJoined{
+			ID:     c.User.ID,
+			Name:   c.User.Name,
+			Avatar: c.User.Avatar,
+		})
+		for _, u := range r.users() {
+			c.Emit(event.UserJoined{
+				ID:     u.ID,
+				Name:   u.Name,
+				Avatar: u.Avatar,
+			})
+		}
+		for _, v := range r.votes {
+			c.Emit(v)
+		}
+	case domain.Finished:
+		r.emitFinish(c)
+	default:
+		return fmt.Errorf("cannot join room in state %s", r.state)
+	}
+
+	c.Emit(event.SettingsUpdate{
+		Location:    r.lobby.Location,
+		PriceMin:    r.lobby.PriceAvg - 300,
+		PriceMax:    r.lobby.PriceAvg + 300,
+		MaxDistance: 4000,
+		Tags: algo.Map(r.lobby.Tags, func(t *domain.Tag) int64 {
+			return t.ID
+		}),
+		RecommendationOpts: r.recommendationOpts,
+	})
 
 	return nil
 }
 
-func (r *Room) AddUser(user *domain.User) error {
-	r.lock.Lock()
-	defer r.lock.Unlock()
-
-	if r.state != domain.InLobby {
-		return fmt.Errorf("can't add user to lobby in state %s", r.state)
-	}
-
-	if _, has := r.usersMap[user.ID]; has {
+func (r *Room) OnLeave(c *state.Context[*Room]) error {
+	_, has := r.usersMap[c.User.ID]
+	if !has {
 		return nil
 	}
-	r.usersMap[user.ID] = user
+	delete(r.usersMap, c.User.ID)
 
-	if r.state == domain.Swiping {
-		r.usersPlace[user.ID] = r.places[0]
+	if r.state == domain.InLobby {
+		err := r.syncUsersWithBd(c.Ctx)
+		if err != nil {
+			return fmt.Errorf("error while syncing users with bd: %w", err)
+		}
 	}
 
-	return r.syncUsersWithBd()
+	c.BroadcastToOthers(event.UserLeft{
+		ID: c.User.ID,
+	})
+
+	return nil
 }
 
-func (r *Room) syncUsersWithBd() error {
+func (r *Room) Empty() bool {
+	r.lock.RLock()
+	defer r.lock.RUnlock()
+	return len(r.usersMap) == 0
+}
+
+func (r *Room) OnLeaveLobby(c *state.Context[*Room], _ event.LeaveLobby) error {
+	err := c.Close()
+	if err != nil {
+		return fmt.Errorf("error while closing connection: %w", err)
+	}
+	return nil
+}
+
+func (r *Room) syncUsersWithBd(ctx context.Context) error {
 	var err error
-	r.lobby.Users, err = r.lobbyUseCase.SetLobbyUsers(
-		context.Background(),
+	r.lobby.Users, err = r.lobbyUseCase.SetLobbyUsers(ctx,
 		r.lobby.ID,
 		algo.Map(r.users(), func(u *domain.User) string {
 			return u.ID
@@ -299,66 +282,26 @@ func (r *Room) syncUsersWithBd() error {
 	return nil
 }
 
-func (r *Room) RemoveUser(id string) error {
+func (r *Room) OnSettingsUpdate(c *state.Context[*Room], ev event.SettingsUpdate) error {
 	r.lock.Lock()
 	defer r.lock.Unlock()
 
-	_, has := r.usersMap[id]
-	if !has {
-		return nil
+	err := r.updateLobbySettings(
+		c.Ctx,
+		ev.Location,
+		(ev.PriceMax+ev.PriceMin)/2,
+		ev.Tags,
+		nil,
+		ev.RecommendationOpts,
+	)
+	if err != nil {
+		return fmt.Errorf("error while updating lobby settings: %w", err)
 	}
-	delete(r.usersMap, id)
 
-	if r.state == domain.InLobby {
-		err := r.syncUsersWithBd()
-		if err != nil {
-			return fmt.Errorf("error while syncing users with bd: %w", err)
-		}
-	}
+	ev.UserID = c.User.ID
+	c.Broadcast(ev)
 
 	return nil
-}
-
-func (r *Room) Empty() bool {
-	r.lock.RLock()
-	defer r.lock.RUnlock()
-	return len(r.usersMap) == 0
-}
-
-func (r *Room) State() domain.LobbyState {
-	r.lock.RLock()
-	defer r.lock.RUnlock()
-	return r.state
-}
-
-func (r *Room) InLobby() bool {
-	r.lock.RLock()
-	defer r.lock.RUnlock()
-	return r.state == domain.InLobby
-}
-
-func (r *Room) Swiping() bool {
-	r.lock.RLock()
-	defer r.lock.RUnlock()
-	return r.state == domain.Swiping
-}
-
-func (r *Room) Finished() bool {
-	r.lock.RLock()
-	defer r.lock.RUnlock()
-	return r.state == domain.Finished
-}
-
-func (r *Room) UpdateLobbySettings(
-	ctx context.Context,
-	location domain.Coordinate,
-	priceAvg int,
-	tagIDs, placeIDs []int64,
-	recommendationOpts *domain.RecommendationOpts,
-) error {
-	r.lock.Lock()
-	defer r.lock.Unlock()
-	return r.updateLobbySettings(ctx, location, priceAvg, tagIDs, placeIDs, recommendationOpts)
 }
 
 func (r *Room) updateLobbySettings(
@@ -385,24 +328,24 @@ func (r *Room) updateLobbySettings(
 	return nil
 }
 
-func (r *Room) StartSwipes(ctx context.Context) error {
+func (r *Room) OnStartSwipes(c *state.Context[*Room], ev event.StartSwipes) error {
 	r.lock.Lock()
 	defer r.lock.Unlock()
 
-	r.log.Debug("Request started: Action 'StartSwipes' initiated")
+	c.Log.Debug("Request started: Action 'StartSwipes' initiated")
 
 	if len(r.lobby.Tags) == 0 {
-		r.log.Warn("Action 'StartSwipes' encountered an issue. Reason: 'No tags found, using default tags'")
-		err := r.updateLobbySettings(ctx, r.lobby.Location, 500, []int64{3}, nil, r.recommendationOpts)
+		c.Log.Warn("Action 'StartSwipes' encountered an issue. Reason: 'No tags found, using default tags'")
+		err := r.updateLobbySettings(c.Ctx, r.lobby.Location, 500, []int64{3}, nil, r.recommendationOpts)
 		if err != nil {
-			r.log.WithError(err).Error("Action 'UpdateLobby' failed")
+			c.Log.WithError(err).Error("Action 'UpdateLobby' failed")
 			return err
 		}
-		r.log.Debug("Request successful: Action 'UpdateLobby' completed with default tags")
+		c.Log.Debug("Request successful: Action 'UpdateLobby' completed with default tags")
 	}
 
 	var err error
-	r.places, err = r.placeRecommender.RecommendPlaces(ctx,
+	r.places, err = r.placeRecommender.RecommendPlaces(c.Ctx,
 		r.recommendationOpts,
 		domain.RecommendData{
 			Location: r.lobby.Location,
@@ -411,12 +354,12 @@ func (r *Room) StartSwipes(ctx context.Context) error {
 		},
 	)
 	if err != nil {
-		r.log.WithError(err).Error("Action 'GetPlacesForLobby' failed")
+		c.Log.WithError(err).Error("Action 'GetPlacesForLobby' failed")
 		return err
 	}
-	r.log.Debug("Request successful: Action 'GetPlacesForLobby' completed")
+	c.Log.Debug("Request successful: Action 'GetPlacesForLobby' completed")
 
-	err = r.updateLobbySettings(ctx, r.lobby.Location, r.lobby.PriceAvg,
+	err = r.updateLobbySettings(c.Ctx, r.lobby.Location, r.lobby.PriceAvg,
 		algo.Map(r.lobby.Tags, func(t *domain.Tag) int64 {
 			return t.ID
 		}),
@@ -428,13 +371,13 @@ func (r *Room) StartSwipes(ctx context.Context) error {
 	if err != nil {
 		return fmt.Errorf("error while updating lobby settings: %w", err)
 	}
-	r.log.Debug("Request successful: Action 'UpdateLobby' completed")
+	c.Log.Debug("Request successful: Action 'UpdateLobby' completed")
 
-	r.log.Debug("Request successful: Action 'UpdateLobby' completed")
+	c.Log.Debug("Request successful: Action 'UpdateLobby' completed")
 	for id := range r.usersMap {
 		if len(r.places) > 0 {
 			r.usersPlace[id] = r.places[0]
-			r.log.Debugf("<User %s> is assigned to <Place %d>", id, r.places[0].ID)
+			c.Log.Debugf("<User %s> is assigned to <Place %d>", id, r.places[0].ID)
 		} else {
 			log.Warnf("No places available to assign to <User %s>", id)
 		}
@@ -442,191 +385,205 @@ func (r *Room) StartSwipes(ctx context.Context) error {
 
 	log.Info("Swipes successfully started")
 	err = r.setState(domain.Swiping)
-	return err
+	if err != nil {
+		return fmt.Errorf("error while setting state: %w", err)
+	}
+
+	c.Broadcast(ev)
+	c.ForEach(func(cc *state.Context[*Room]) {
+		p := r.usersPlace[cc.User.ID]
+		cc.Emit(event.Place{
+			ID:   p.ID,
+			Card: p,
+		})
+	})
+
+	return nil
 }
 
-func (r *Room) Swipe(userID string, placeID int64, t domain.SwipeType) (*MatchVote, error) {
+func (r *Room) OnSwipe(c *state.Context[*Room], ev event.Swipe) error {
 	r.lock.Lock()
 	defer r.lock.Unlock()
 
-	r.swipes = append(r.swipes, &domain.Swipe{
+	placeID := r.usersPlace[c.User.ID].ID
+
+	swipe := &domain.Swipe{
 		LobbyID: r.lobby.ID,
 		PlaceID: placeID,
-		UserID:  userID,
-		Type:    t,
-	})
-	err := r.swipeUseCase.SaveSwipe(context.Background(), &domain.Swipe{
-		LobbyID: r.lobby.ID,
-		PlaceID: placeID,
-		UserID:  userID,
-		Type:    t,
-	})
+		UserID:  c.User.ID,
+		Type:    ev.SwipeType,
+	}
+
+	r.swipes = append(r.swipes, swipe)
+	err := r.swipeUseCase.SaveSwipe(context.Background(), swipe)
 	if err != nil {
-		return nil, fmt.Errorf("error while saving swipe: %w", err)
+		return fmt.Errorf("error while saving swipe: %w", err)
 	}
 
 	pIdx := slices.IndexFunc(r.places, func(place *domain.Place) bool {
 		return place.ID == placeID
 	})
-	r.usersPlace[userID] = r.places[(pIdx+1)%len(r.places)]
+	r.usersPlace[c.User.ID] = r.places[(pIdx+1)%len(r.places)]
 
 	likes := algo.Count(r.swipes, func(swipe *domain.Swipe) bool {
 		return swipe.PlaceID == placeID && swipe.Type == domain.LIKE
 	})
+
 	if likes == len(r.usersMap) {
-		match := &Match{Place: r.places[slices.IndexFunc(r.places, func(place *domain.Place) bool {
+		match := &event.Match{Place: r.places[slices.IndexFunc(r.places, func(place *domain.Place) bool {
 			return place.ID == placeID
 		})]}
 		match.ID = len(r.matches)
 		r.matches = append(r.matches, match)
 		r.result = match.Place
 
-		vote := &MatchVote{
-			BaseVote: BaseVote{
+		vote := event.MatchVote{
+			BaseVote: event.BaseVote{
 				ID: int64(len(r.votes)),
-				Options: []VoteOption{
-					{ID: OptionIDLike, Desc: "Like"},
-					{ID: OptionIDDislike, Desc: "Dislike"},
+				Options: []event.VoteOption{
+					{ID: event.OptionIDLike, Desc: "Like"},
+					{ID: event.OptionIDDislike, Desc: "Dislike"},
 				},
-				Type:  VoteTypeMatch,
-				votes: make(map[string]OptionID),
+				Type:  event.VoteTypeMatch,
+				Votes: make(map[string]event.OptionID),
 			},
 			Place: match.Place,
 		}
 		r.votes[vote.ID] = vote
-		return vote, nil
+		c.Broadcast(vote)
 	}
 
-	return nil, nil
-}
+	c.Emit(event.Place{
+		ID:   r.usersPlace[c.User.ID].ID,
+		Card: r.usersPlace[c.User.ID],
+	})
 
-func (r *Room) setState(state domain.LobbyState) error {
-	r.state = state
-	err := r.lobbyUseCase.SetLobbyState(context.Background(), r.lobby.ID, state)
-	return err
-}
-
-func (r *Room) Result() *domain.Place {
-	r.lock.RLock()
-	defer r.lock.RUnlock()
-	if r.result != nil {
-		return r.result
-	}
-	if len(r.matches) > 0 {
-		return r.matches[len(r.matches)-1].Place
-	}
 	return nil
 }
 
-func (r *Room) Matches() []*Match {
-	r.lock.RLock()
-	defer r.lock.RUnlock()
-	return r.matches
-}
-
-func (r *Room) Votes() []Vote {
-	r.lock.RLock()
-	defer r.lock.RUnlock()
-	votes := make([]Vote, 0, len(r.votes))
-	for _, vote := range r.votes {
-		votes = append(votes, vote)
-	}
-	return votes
-}
-
-func (r *Room) Vote(userID string, voteID int64, optionID OptionID) (*VoteResult, error) {
+func (r *Room) OnVote(c *state.Context[*Room], ev event.Vote) error {
 	r.lock.Lock()
 	defer r.lock.Unlock()
 
-	vote, ok := r.votes[voteID]
+	vote, ok := r.votes[ev.VoteID]
 	if !ok {
-		return nil, fmt.Errorf("vote with id %d not found", voteID)
+		return fmt.Errorf("vote with id %d not found", ev.VoteID)
 	}
 
+	var res *event.VoteResult
+
 	switch v := vote.(type) {
-	case *MatchVote:
-		if optionID != OptionIDLike && optionID != OptionIDDislike {
-			return nil, fmt.Errorf("invalid option id: %d", optionID)
+	case event.MatchVote:
+		if ev.OptionID != event.OptionIDLike && ev.OptionID != event.OptionIDDislike {
+			return fmt.Errorf("invalid option id: %d", ev.OptionID)
 		}
-		v.votes[userID] = optionID
+		v.Votes[c.User.ID] = ev.OptionID
 
-		log.Debugf("<User %s> voted for <Option %d> in %s", userID, optionID, v.Type)
-		res := r.voteMatchResult(v)
+		log.Debugf("<User %s> voted for <Option %d> in %s", c.User.ID, ev.OptionID, v.Type)
+		res = r.voteMatchResult(v)
 
-		if res != nil && res.OptionID == OptionIDLike {
+		if res != nil && res.OptionID == event.OptionIDLike {
 			r.result = v.Place
 			err := r.setState(domain.Finished)
 			if err != nil {
-				return nil, fmt.Errorf("error while setting lobby state: %w", err)
+				return fmt.Errorf("error while setting lobby state: %w", err)
 			}
-			return res, nil
 		}
 
-		return res, nil
-	case *FinishVote:
-		if optionID != OptionIDFinish && optionID != OptionIDContinue {
-			return nil, fmt.Errorf("invalid option id: %d", optionID)
+	case event.FinishVote:
+		if ev.OptionID != event.OptionIDFinish && ev.OptionID != event.OptionIDContinue {
+			return fmt.Errorf("invalid option id: %d", ev.OptionID)
 		}
-		v.votes[userID] = optionID
+		v.Votes[c.User.ID] = ev.OptionID
 
-		log.Debugf("<User %s> voted for <Option %d> in %s", userID, optionID, v.Type)
-		res := r.voteFinishResult(v)
-		if res != nil && res.OptionID == OptionIDFinish {
+		log.Debugf("<User %s> voted for <Option %d> in %s", c.User.ID, ev.OptionID, v.Type)
+		res = r.voteFinishResult(v)
+		if res != nil && res.OptionID == event.OptionIDFinish {
 			err := r.setState(domain.Finished)
 			if err != nil {
-				return nil, fmt.Errorf("error while setting lobby state: %w", err)
+				return fmt.Errorf("error while setting lobby state: %w", err)
 			}
-			return res, nil
 		}
 
-		return res, nil
+	default:
+		log.Warnf("Unknown voting type: %T", vote)
+		return fmt.Errorf("unknown voting type: %T", vote)
 	}
 
-	log.Warnf("Unknown voting type: %T", vote)
-	return nil, fmt.Errorf("unknown voting type: %T", vote)
+	c.Broadcast(event.Voted{
+		VoteID:   ev.VoteID,
+		OptionID: ev.OptionID,
+		User: event.UserJoined{
+			ID:     c.User.ID,
+			Name:   c.User.Name,
+			Avatar: c.User.Avatar,
+		},
+	})
+
+	if res != nil {
+		c.Broadcast(res)
+	}
+
+	if r.state == domain.Finished {
+		c.ForEach(r.emitFinish)
+	}
+
+	return nil
 }
 
-func (r *Room) voteMatchResult(vote *MatchVote) *VoteResult {
-	if len(vote.votes) != len(r.usersMap) {
+func (r *Room) voteMatchResult(vote event.MatchVote) *event.VoteResult {
+	if len(vote.Votes) != len(r.usersMap) {
 		return nil
 	}
 	likes := 0
-	for _, vote := range vote.votes {
-		if vote == OptionIDLike {
+	for _, vote := range vote.Votes {
+		if vote == event.OptionIDLike {
 			likes++
 		}
 	}
 	if likes == len(r.usersMap) {
-		return &VoteResult{
+		return &event.VoteResult{
 			VoteID:   vote.ID,
-			Type:     VoteTypeMatch,
-			OptionID: OptionIDLike,
+			Type:     event.VoteTypeMatch,
+			OptionID: event.OptionIDLike,
 		}
 	}
 
-	return &VoteResult{
+	return &event.VoteResult{
 		VoteID:   vote.ID,
-		Type:     VoteTypeMatch,
-		OptionID: OptionIDDislike,
+		Type:     event.VoteTypeMatch,
+		OptionID: event.OptionIDDislike,
 	}
 }
 
-func (r *Room) voteFinishResult(vote *FinishVote) *VoteResult {
-	if len(vote.votes) != len(r.usersMap) {
+func (r *Room) voteFinishResult(vote event.FinishVote) *event.VoteResult {
+	if len(vote.Votes) != len(r.usersMap) {
 		return nil
 	}
 	likes := 0
-	for _, vote := range vote.votes {
-		if vote == OptionIDFinish {
+	for _, vote := range vote.Votes {
+		if vote == event.OptionIDFinish {
 			likes++
 		}
 	}
 	if likes == len(r.usersMap) {
-		return &VoteResult{
+		return &event.VoteResult{
 			VoteID:   vote.ID,
-			Type:     VoteTypeFinish,
-			OptionID: OptionIDFinish,
+			Type:     event.VoteTypeFinish,
+			OptionID: event.OptionIDFinish,
 		}
 	}
 	return nil
+}
+
+func (r *Room) emitFinish(c *state.Context[*Room]) {
+	var matchResult *domain.Place
+	if len(r.matches) > 0 {
+		matchResult = r.matches[len(r.matches)-1].Place
+	}
+
+	c.Emit(event.Finish{
+		Result:  matchResult,
+		Matches: r.matches,
+	})
 }
